Move parsing logic into helper
